pkg/release: move from pkg/framework/resource (#137)
