+++ pkg/release/release.go
@@ -1,4 +1,4 @@
-package resource
+package release
 
 import (
 	"context"
